Add files via upload
Small formatting changes. ':' on the right side of the column label denotes align right. This is much faster than testing for number like data.
Table is responsible for filling in missing spaces now.
Table and align both have a cast parameter which is just a single argument function to convert the data from/to dictionary formats as needed.

Removed matrix
Updated themes a bit.